--- Assets/AddOn/GlitchFx-master/GlitchFx.unitypackage.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: d80f995af690a474e841364c1e88ee1b
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
