+++ Assets/CSV/BT15.csv
@@ -4,7 +4,7 @@ BT15-001_P1,Yokomon_BT15_001,3115,,,,,,
 BT15-002,Tsunomon_BT15_002,3116,,,,,,
 BT15-002_P1,Tsunomon_BT15_002,3117,,,,,,
 BT15-003,Nyaromon_BT15_003,3118,,,,,,
-BT15-003_P1,Nyaromon_BT15_004,3119,,,,,,
+BT15-003_P1,Nyaromon_BT15_003,3119,,,,,,
 BT15-004,Motimon_BT15_004,3120,,,,,,
 BT15-004_P1,Motimon_BT15_004,3121,,,,,,
 BT15-005,Kyokyomon_BT15_005,3122,,,,,,
