--- Assets/CardEffect/BT15/Red/DualWingOfLove_BT15_088.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: fac1ee684717187419754b68b7c929fc
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
